refactored code and added validation to cite new url form
